requirements.txt
paho-mqtt
asyncio-mqtt
pyyaml